# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/profile": {
      "post": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findProfile",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/profile/change-password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/sales": {
      "post": {
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "get": {
        "operationId": "SalesController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "customerName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SaleRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/resports": {
      "get": {
        "operationId": "SalesController_generateReport",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "customerName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/{id}": {
      "get": {
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "patch": {
        "operationId": "SalesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "delete": {
        "operationId": "SalesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/resport/{id}": {
      "get": {
        "operationId": "SalesController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
...
  },
  "info": {
    "title": "MI PROFE",
    "description": "The MI PROFE-API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "role",
          "idDocument",
          "lastName",
          "firstName",
          "status",
          "birthdate"
        ]
      },
      "UserRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lastName",
          "firstName",
          "email",
          "role",
          "idDocument",
          "status",
          "birthdate"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "birthdate": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "ChangePasswordUserDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "confirmPassword"
        ]
      },
...
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        }
      },
...
      "CreateSaleProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "subtotal": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "amount",
          "price",
          "subtotal",
          "product"
        ]
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        },
        "required": [
          "date",
          "stage",
          "total",
          "customer",
          "saleProducts"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {}
      },
      "SaleRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        },
        "required": [
          "id",
          "date",
          "stage",
          "total",
          "customer",
          "saleProducts"
        ]
      },
      "ReportsResponseDto": {
        "type": "object",
        "properties": {
          "reportUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "reportUrl",
          "name"
        ]
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        }
      },
...
    }
  }
}